refactor: split board storage from import, rename boards-db.ts
Storage holds persistence only (db, queries, board-sets store, hooks).
Import owns the OBF/OBZ → DB write path plus URL/file-picker entrypoints,
and depends on storage one-way. The wrapped importBoardFiles (write +
notify + broadcast) moves into import/; storage/board-sets-store exposes
notifyBoardSetsChanged for it. writeBoardSetFiles is the pure write,
exported for tests so they don't mutate the singleton store.
